Add rule to enforce consuming only topics in current module
This is rather tied to the expected implementation of defining
consumes/producers using our `tls-app` module[1], specifically a
`module` block that uses at least one of `consume_topics` or
`produce_topics`.

`tflint` doesn't appear to do much (any) processing before running, so
e.g.

* Doesn't populate bits following a `source = ./path/to/source`
* Doesn't do any variable expansion

A poke around of some existing linters gives the impression that the
approach generally taken is just to skip running checks on these cases
(e.g. [2] is a change is to a linter that runs on AWS resource tags, but
just skips any tag that isn't a string literal. But, since almost all
our produces/consumers are defined in blocks like:

    module "example_producer" {
      source           = "../../../modules/tls-app"
      produce_topics   = [kafka_topic.pubsub_examples.name]
      cert_common_name = "pubsub/example-producer"
    }

Simply skipping the variable name (`kafka_topic.pubsub_examples.name`
in the example) would essentially mean the linter would skip every
block. So the linter checks on `module`s of that structure.

Another approach I looked at was to use the `Expression.Variables`[3] and
then try and build up a name from the parts of that, but I couldn't find
a particularly clean way to do that.

[1] https://github.com/utilitywarehouse/kafka-cluster-config/tree/d28b50424f9d38d31ddceb50338df62a6934694f/modules/tls-app
[2] https://github.com/terraform-linters/tflint-ruleset-aws/commit/666253c1189a75fc3b8a46c5ed24cd59992a7e49
[3] https://pkg.go.dev/github.com/hashicorp/hcl/v2#Expression

Ticket: DENA-975