Fix panic on non-string topic name
A panic results in an unhelpful message when running the plugin:

    Failed to check ruleset; error reading from server: EOF

So report an error here instead.

There is potential for similar behaviour if `consume_topics` is not a
list, in which case `AsValueSlice` would panic, but I'm content to not
be too defensive here if no-ones needing that change.

Ticket: DENA-1175